Use v1beta1.LiteralFloat32

+++ controllers/activemqartemis_controller.go
@@ -251,15 +251,15 @@ func UpdateCR(cr *brokerv1beta1.ActiveMQArtemis, client rtclient.Client, namespa
 	// in v2alpha5/v0.20.1 and they are *string in v2alpha5/v1.0.0.
 	// Those fields has been reverted to the original type to fix this backward compatibility issue
 	// but this reversion has caused a conversion issue from v2alpha5 to v1beta1.
-	// To fix this conversion issue the type of v1beta1 fields have been changed to json.Number.
-	// The json.Number allows to unmarshal float and string values but it only allows to marshal string values.
+	// To fix this conversion issue the type of v1beta1 fields have been changed to v1beta1.LiteralFloat32.
+	// The v1beta1.LiteralFloat32 allows to unmarshal float and string values but it only allows to marshal string values.
 	// The following code block patches CR that contain float values and uses the string lenght to check
 	// if those values has been already patched.
 	patchPayload := []map[string]string{}
 	addressSettings := cr.Spec.AddressSettings.AddressSetting
 	for i := 0; i < len(addressSettings); i++ {
 		if addressSettings[i].RedeliveryDelayMultiplier != nil && len(addressSettings[i].RedeliveryDelayMultiplier.String()) < 9 {
-			redeliveryDelayMultiplier, err := addressSettings[i].RedeliveryDelayMultiplier.Float64()
+			redeliveryDelayMultiplier, err := addressSettings[i].RedeliveryDelayMultiplier.Float()
 			if err != nil {
 				clog.Error(err, "Error parsing RedeliveryDelayMultiplier", "ActiveMQArtemis", namespacedName)
 				return err
@@ -271,7 +271,7 @@ func UpdateCR(cr *brokerv1beta1.ActiveMQArtemis, client rtclient.Client, namespa
 			})
 		}
 		if addressSettings[i].RedeliveryCollisionAvoidanceFactor != nil && len(addressSettings[i].RedeliveryCollisionAvoidanceFactor.String()) < 9 {
-			redeliveryCollisionAvoidanceFactor, err := addressSettings[i].RedeliveryCollisionAvoidanceFactor.Float64()
+			redeliveryCollisionAvoidanceFactor, err := addressSettings[i].RedeliveryCollisionAvoidanceFactor.Float()
 			if err != nil {
 				clog.Error(err, "Error parsing RedeliveryCollisionAvoidanceFactor", "ActiveMQArtemis", namespacedName)
 				return err
